Added a utility method to take a protobuf Enum and return the 'friendly name' of it (given the reflection object and the enum type name), also did the same treatment i did with server_database.py where technically, the data_received() method can be called with bytes, but it not might be the complete message, so i added functionality to read the first 2 bytes to get the size of the rest of the message, and then if we don't get SIZE bytes in the rest of the data, then we save the state and resume on the next data_received() call. This is more complicated then the server_database.py implementation (_sockRecv/Send) because the only way we get data is when asyncio/Protocol calls our data_received method, so we have to save state in the class and whatnot.
Because of that, i separated data_received into reading the data and putting together the chunks if necessary , and then once the message is complete we we call complete_data_received().

also made it so that we append the data that we send to the client with the 2 byte size prefix.

-----------------
exported from bzr_postcommit_to_git.py,
going from bzr commit:
82:markgrandi@gmail.com-20140904230513-xs56t4wtt62969jr =>
83:markgrandi@gmail.com-20140904230858-mb5tyidu3tva0ohs